LPD-68504 Remove util's dependency on test class so other test classes can use it too.
